Crate modular_bitfield_impl

Macros

define_specifiers

Generates the B1, B2, …, B128 bitfield specifiers.

Attribute Macros

bitfield

Applicable to structs to turn their fields into compact bitfields.

Derive Macros

BitfieldSpecifier

Derive macro for Rust enums to implement Specifier trait.